backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-api-python-client>=2.0.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pydantic>=1.8.0
python-multipart>=0.0.5
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3
pytest>=6.2.5
pytest-asyncio>=0.16.0
httpx>=0.24.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
python-jose>=3.3.0
passlib>=1.7.4
tweepy==4.14.0
schedule==1.2.1
python-dateutil>=2.8.2
openai==1.12.0
pytz>=2021.3

# テスト用の依存関係
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
mypy>=0.910
flake8>=3.9.2
textblob==0.17.1
numpy==1.26.4
google-generativeai==0.3.2 